Solved: 2149 of LeetCode

+++ 4_Maths/Combination_ncr.cpp
@@ -1,13 +1,13 @@
 #include <iostream>
 using namespace std;
-int factorial (int n) {
-    int ans = 1;
+long long factorial (int n) {
+    long long ans = 1;
     for (int i = 1; i <= n; i++) {
         ans *= i;
     }
     return ans;
 }
-int nCr(int n, int r) {
+long long nCr(int n, int r) {
     return factorial(n) / (factorial(r) * factorial(n-r));
 }
 int main() {
